==> Klargøre data/2 udlede muslimske typologi.Rmd <==
---
title: "2 Udlede muslimske typologi"
output: html_notebook

editor_options: 
  markdown: 
    wrap: 72
header-includes:
    - \usepackage{setspace}\doublespacing
    - \usepackage{booktabs} 
    - \usepackage{tabularx}
    - \usepackage{tidy}
---

## Åbne data

```{r åbne data, message=FALSE, include=FALSE, echo=FALSE}
ess_immigration <- read.csv("../data/ess_all.csv")
```

## Loaded relevante pakker

```{r hente-pakker, message=FALSE, include=FALSE, echo=FALSE}
library(tidyverse)
#Sys.setenv("LANGUAGE"="EN") #error in english
start.time <- Sys.time()
```

## Oprindelsesland

Variable som dækker over land som respondent og respondent forældre er
født i hedder noget forskelligt påtværs af ESS runder. Derfor samler jeg
det i tre variabler

### respondents oprindelsesland

```{r}
ess_cntrbrth <- ess_immigration%>%
  dplyr::select(id, essround, brncntr, cntbrth, cntbrtha, cntbrthb, cntbrthc, cntbrthd)%>%
  filter(brncntr == 2)
```

```{r}
ess_cntrbrth$cntbrth_1[is.na(ess_cntrbrth$cntbrtha) & is.na(ess_cntrbrth$cntbrthb) & is.na(ess_cntrbrth$cntbrthc) & is.na(ess_cntrbrth$cntbrthd)] <- ess_cntrbrth$cntbrth[is.na(ess_cntrbrth$cntbrtha) & is.na(ess_cntrbrth$cntbrthb) & is.na(ess_cntrbrth$cntbrthc) & is.na(ess_cntrbrth$cntbrthd)]

ess_cntrbrth$cntbrth_1[is.na(ess_cntrbrth$cntbrth) & is.na(ess_cntrbrth$cntbrthb) & is.na(ess_cntrbrth$cntbrthc) & is.na(ess_cntrbrth$cntbrthd)] <- ess_cntrbrth$cntbrtha[is.na(ess_cntrbrth$cntbrth) & is.na(ess_cntrbrth$cntbrthb) & is.na(ess_cntrbrth$cntbrthc) & is.na(ess_cntrbrth$cntbrthd)]

ess_cntrbrth$cntbrth_1[is.na(ess_cntrbrth$cntbrtha) & is.na(ess_cntrbrth$cntbrth) & is.na(ess_cntrbrth$cntbrthc) & is.na(ess_cntrbrth$cntbrthd)] <- ess_cntrbrth$cntbrthb[is.na(ess_cntrbrth$cntbrth) & is.na(ess_cntrbrth$cntbrtha) & is.na(ess_cntrbrth$cntbrthc) & is.na(ess_cntrbrth$cntbrthd)]

ess_cntrbrth$cntbrth_1[is.na(ess_cntrbrth$cntbrtha) & is.na(ess_cntrbrth$cntbrthb) & is.na(ess_cntrbrth$cntbrth) & is.na(ess_cntrbrth$cntbrthd)] <- ess_cntrbrth$cntbrthc[is.na(ess_cntrbrth$cntbrtha) & is.na(ess_cntrbrth$cntbrthb) & is.na(ess_cntrbrth$cntbrth) & is.na(ess_cntrbrth$cntbrthd)]

ess_cntrbrth$cntbrth_1[is.na(ess_cntrbrth$cntbrtha) & is.na(ess_cntrbrth$cntbrthb) & is.na(ess_cntrbrth$cntbrthc) & is.na(ess_cntrbrth$cntbrth)] <- ess_cntrbrth$cntbrthd[is.na(ess_cntrbrth$cntbrtha) & is.na(ess_cntrbrth$cntbrthb) & is.na(ess_cntrbrth$cntbrthc) & is.na(ess_cntrbrth$cntbrth)]
```

```{r}
ess_immigration <- ess_immigration %>%
  dplyr::select(-c("cntbrth", "cntbrtha", "cntbrthb", "cntbrthc", "cntbrthd"))

ess_cntrbrth <- ess_cntrbrth %>%
  dplyr::select(-c("cntbrth", "cntbrtha", "cntbrthb", "cntbrthc", "cntbrthd", "essround", "brncntr"))
```

```{r}
ess_immigration <- left_join(ess_immigration, ess_cntrbrth, by = "id")
rm(ess_cntrbrth)
```

```{r}
ess_immigration <- ess_immigration %>%
  rename(cntbrth = cntbrth_1)
```

## forældres oprindelsesland

Variable som dækker over land som respondent er født i hedder noget
forskelligt påtværs af ESS runder. Jeg samler det i en variable:
cntrbrth Far og mor land født i far født i landet: facntr mor født i
landet: mocntr 1: ja, 2: nej

sprøger ikke om hvilket land født i ess 1-2 land født i far: fbrncnt
(3), fbrncnta (4-5), fbrncntb (6-8), fbrncntc (9) land født i mor:
mbrncnt (3), mbrncnta (4-5), mbrncntb (6-8), mbrncntc (9)

### Fars oprindelsesland

```{r}
ess_far <- ess_immigration %>%
  dplyr::select(id, essround, facntr, fbrncnt, fbrncnta, fbrncntb, fbrncntc)%>%
  filter(facntr == 2)
```

```{r}
ess_far$fbrncnt_1[is.na(ess_far$fbrncnta) & is.na(ess_far$fbrncntb) & is.na(ess_far$fbrncntc)] <- ess_far$fbrncnt[is.na(ess_far$fbrncnta) & is.na(ess_far$fbrncntb) & is.na(ess_far$fbrncntc)]

ess_far$fbrncnt_1[is.na(ess_far$fbrncnt) & is.na(ess_far$fbrncntb) & is.na(ess_far$fbrncntc)] <- ess_far$fbrncnta[is.na(ess_far$fbrncnt) & is.na(ess_far$fbrncntb) & is.na(ess_far$fbrncntc)]

ess_far$fbrncnt_1[is.na(ess_far$fbrncnta) & is.na(ess_far$fbrncnt) & is.na(ess_far$fbrncntc)] <- ess_far$fbrncntb[is.na(ess_far$fbrncnt) & is.na(ess_far$fbrncnta) & is.na(ess_far$fbrncntc)]

ess_far$fbrncnt_1[is.na(ess_far$fbrncnta) & is.na(ess_far$fbrncntb) & is.na(ess_far$fbrncnt)] <- ess_far$fbrncntc[is.na(ess_far$fbrncnta) & is.na(ess_far$fbrncntb) & is.na(ess_far$fbrncnt)]

```

```{r}
ess_immigration <- ess_immigration %>%
  dplyr::select(-c("fbrncnta", "fbrncntb", "fbrncntc", "fbrncnt"))

ess_far <- ess_far %>%
  dplyr::select(-c("fbrncnta", "fbrncntb", "fbrncntc", "fbrncnt", "essround", "facntr"))

ess_immigration <- left_join(ess_immigration, ess_far, by = "id")
```

```{r}
ess_immigration <- ess_immigration %>%
  rename(fbrncnt = fbrncnt_1)

rm(ess_far)
```

### Mors oprindelsesland

```{r}
ess_mor <- ess_immigration %>%
  dplyr::select(id, essround, mocntr, mbrncnt, mbrncnta, mbrncntb, mbrncntc)%>%
  filter(mocntr == 2)
```

```{r}
ess_mor$mbrncnt_1[is.na(ess_mor$mbrncnta) & is.na(ess_mor$mbrncntb) & is.na(ess_mor$mbrncntc)] <- ess_mor$mbrncnt[is.na(ess_mor$mbrncnta) & is.na(ess_mor$mbrncntb) & is.na(ess_mor$mbrncntc)]

ess_mor$mbrncnt_1[is.na(ess_mor$mbrncnt) & is.na(ess_mor$mbrncntb) & is.na(ess_mor$mbrncntc)] <- ess_mor$mbrncnta[is.na(ess_mor$mbrncnt) & is.na(ess_mor$mbrncntb) & is.na(ess_mor$mbrncntc)]

ess_mor$mbrncnt_1[is.na(ess_mor$mbrncnta) & is.na(ess_mor$mbrncnt) & is.na(ess_mor$mbrncntc)] <- ess_mor$mbrncntb[is.na(ess_mor$mbrncnt) & is.na(ess_mor$mbrncnta) & is.na(ess_mor$mbrncntc)]

ess_mor$mbrncnt_1[is.na(ess_mor$mbrncnta) & is.na(ess_mor$mbrncntb) & is.na(ess_mor$mbrncnt)] <- ess_mor$mbrncntc[is.na(ess_mor$mbrncnta) & is.na(ess_mor$mbrncntb) & is.na(ess_mor$mbrncnt)]

```

```{r}
ess_immigration <- ess_immigration %>%
  dplyr::select(-c("mbrncnta", "mbrncntb", "mbrncntc", "mbrncnt"))

ess_mor <- ess_mor %>%
  dplyr::select(-c("mbrncnta", "mbrncntb", "mbrncntc", "mbrncnt", "essround", "mocntr"))

ess_immigration <- left_join(ess_immigration, ess_mor, by = "id")
```

```{r}
ess_immigration <- ess_immigration %>%
  rename(mbrncnt = mbrncnt_1)

rm(ess_mor)
```

## Antal år levet i ESS land

bruges ikke til at udlede muslimsk identitet

```{r}
ess_live <- ess_immigration%>%
  dplyr::select(id, essround, livecntr, int.year, livecnta)%>%
  mutate(years = int.year-livecnta)

ess_live$livecntr[ess_live$years == 0] <- 1 
ess_live$livecntr[ess_live$years %in% c(1,2,3,4,5)] <- 2
ess_live$livecntr[ess_live$years %in% c(6:10)] <- 3
ess_live$livecntr[ess_live$years %in% c(11:20)] <- 4
ess_live$livecntr[ess_live$years %in% c(21:102)] <- 5



ess_live <- ess_live %>%
  dplyr::select(-c(years, int.year, essround))

ess_immigration <- ess_immigration %>%
  dplyr::select(-c(livecntr, livecnta))

ess_immigration <- left_join(ess_immigration, ess_live, by = "id")
rm(ess_live)
```

## Muslimske majoritets lande liste

```{r}
country <- c("MV", "MR", "SO", "TN", "AF", "IR", "EH", "YE", "DZ", "MA", "KM", "NE", "SA", "TJ", "PS", "JO", "DJ", "LY", "YT", "SD", "AZ", "PK", "SN", "GM", "IQ", "XK", "ML", "TM", "BD", "EG", "TR", "GN", "UZ", "SY", "ID", "OM", "KG", "BN", "SL", "QA", "AE", "KW", "BH", "KZ", "BF", "MY", "TD", "LB", "NG", "AL", "BA")
```

## Citizenship

bliver ikke brugt til at udlede muslimsk identitet

```{r}
citizenship <- ess_immigration %>%
  dplyr::select(id, ctzcntr, ctzship, ctzshipa, ctzshipb, ctzshipc, ctzshipd, cntry)

error_ctzship <- c("65", "66", "77", "88", "99", "6500", "6666", "7777", "8888", "9999")


```

```{r}
citizenship$ctzship_1[is.na(citizenship$ctzshipa) & is.na(citizenship$ctzshipb) & is.na(citizenship$ctzshipc) & is.na(citizenship$ctzshipd)] <- citizenship$ctzship[is.na(citizenship$ctzshipa) & is.na(citizenship$ctzshipb) & is.na(citizenship$ctzshipc) & is.na(citizenship$ctzshipd)]

citizenship$ctzship_1[is.na(citizenship$ctzship) & is.na(citizenship$ctzshipb) & is.na(citizenship$ctzshipc) & is.na(citizenship$ctzshipd)] <- citizenship$ctzshipa[is.na(citizenship$ctzship) & is.na(citizenship$ctzshipb) & is.na(citizenship$ctzshipc) & is.na(citizenship$ctzshipd)]

citizenship$ctzship_1[is.na(citizenship$ctzshipa) & is.na(citizenship$ctzship) & is.na(citizenship$ctzshipc) & is.na(citizenship$ctzshipd)] <- citizenship$ctzshipb[is.na(citizenship$ctzship) & is.na(citizenship$ctzshipa) & is.na(citizenship$ctzshipc) & is.na(citizenship$ctzshipd)]

citizenship$ctzship_1[is.na(citizenship$ctzshipa) & is.na(citizenship$ctzshipb) & is.na(citizenship$ctzship) & is.na(citizenship$ctzshipd)] <- citizenship$ctzshipc[is.na(citizenship$ctzshipa) & is.na(citizenship$ctzshipb) & is.na(citizenship$ctzship) & is.na(citizenship$ctzshipd)]

citizenship$ctzship_1[is.na(citizenship$ctzshipa) & is.na(citizenship$ctzshipb) & is.na(citizenship$ctzshipc) & is.na(citizenship$ctzship)] <- citizenship$ctzshipd[is.na(citizenship$ctzshipa) & is.na(citizenship$ctzshipb) & is.na(citizenship$ctzshipc) & is.na(citizenship$ctzship)]

citizenship$ctzship_1[citizenship$ctzship_1 %in% error_ctzship] <- "NA"
```

```{r}
citizenship <- citizenship %>%
  mutate(ctzship_m = ifelse(ctzship_1 == "NA", "NA",
           ifelse(ctzship_1 %in% country, "muslim country", "not muslim country")))
```

```{r}
citizenship <- citizenship %>%
  dplyr::select(id, ctzship_m, ctzship_1)%>%
  rename(ctzship = ctzship_1)

ess_immigration <- ess_immigration %>%
  dplyr::select(-c(ctzship, ctzshipa, ctzshipb, ctzshipc, ctzshipd))

ess_immigration <- left_join(ess_immigration, citizenship, by = "id")
rm(citizenship)
```

## Varialble om fra muslimsk majoritets land, eller forældre

```{r}
ess_immigration <- ess_immigration %>%
  mutate(cntbrth_m = ifelse(cntbrth == "NA", "NA",
           ifelse(cntbrth %in% country, "muslim country", "not muslim country")))


ess_immigration <- ess_immigration %>%
  mutate(mbrncnt_m = ifelse(cntbrth_m == "muslim country", "NA",
                            ifelse(mbrncnt %in% country, "muslim country", "not muslim country")))%>%
  mutate(fbrncnt_m = ifelse(cntbrth_m == "muslim country", "NA",
         ifelse(fbrncnt %in% country, "muslim country", "not muslim country")))
```

## Religion variable

```{r}
religion <- ess_immigration %>%
  dplyr::select(id, rlgblg, rlgdnm, rlgblge, rlgdnme)
```

```{r}
religion$rlgblg[is.na(religion$rlgdnm)] <- 2
religion$rlgblge[is.na(religion$rlgdnme)] <- 2
```

```{r}
religion$religion_now[religion$rlgdnm == 6] <- "muslim"
religion$religion_now[religion$rlgblg == 1 & religion$rlgdnm %in% c(1, 2, 3, 4, 5, 7, 8)] <- "not muslim"
religion$religion_now[religion$rlgblg != 1] <- "no religion"

religion$religion_before[religion$rlgblge == 1 & religion$rlgdnme %in% c(1, 2, 3, 4, 5, 7, 8)]<- "not muslim"
religion$religion_before[religion$rlgblge == 1 & religion$rlgdnme %in% c(6)]<- "muslim"

#religion$religion_before[is.na(religion$rlgdnme)] <- "no religion"

```

```{r}
ess_immigration <- ess_immigration %>%
  dplyr::select(-c(rlgblg, rlgdnm, rlgblge, rlgdnme))

ess_immigration <- left_join(ess_immigration, religion, by = "id")
rm(religion)
```

```{r}
muslim <- ess_immigration %>%
  dplyr::select(id, fbrncnt_m, mbrncnt_m, cntbrth_m, religion_now, religion_before)

```

laves efter at tidligere religion islam - del af respondenter (koder
disse som 2) ingen religion kun en forældre fra muslimsk majoritets land
(kodet 3)

# Udlede muslimsk identitet efter typologi

```{r}
muslim <- muslim %>%
  mutate(
    respondent = case_when(
      #religion
      (religion_now == "not muslim") ~ 0,
      (religion_now == "muslim") ~ 1,
      
      #tidligere religion
      (religion_now == "no religion" &  religion_before == "not muslim") ~ 0,
      (religion_now == "no religion" &  religion_before == "muslim") ~ 1,
      
      
      #oprindelsesland
      (religion_now == "no religion" &  religion_before == "no religion" & 
       cntbrth_m == "muslim country") ~ 2,
      (religion_now == "no religion" &  is.na(religion_before) & 
       cntbrth_m == "muslim country") ~ 2,
      
      
      #forældres oprindelsesland
      (religion_now == "no religion" &  religion_before == "no religion" &
       mbrncnt_m == "muslim country" & fbrncnt_m == "muslim country") ~ 2,
     (religion_now == "no religion" &  is.na(religion_before) &
       mbrncnt_m == "muslim country" & fbrncnt_m == "muslim country") ~ 2,
     (is.na(religion_now) &  is.na(religion_before) &
       mbrncnt_m == "muslim country" & fbrncnt_m == "muslim country") ~ 2,

      (religion_now == "no religion" &  religion_before == "no religion" &
       fbrncnt_m == "not muslim country" & mbrncnt_m == "not muslim country") ~ 0,
      (religion_now == "no religion" &  is.na(religion_before) &
       fbrncnt_m == "not muslim country" & mbrncnt_m == "not muslim country") ~ 0,
     
     #kun den ene forældre fra muslimsk land
      (religion_now == "no religion" &  religion_before == "no religion" &
       mbrncnt_m == "muslim country" & fbrncnt_m != "muslim country") ~ 3,
     (religion_now == "no religion" &  is.na(religion_before) &
       mbrncnt_m == "muslim country" & fbrncnt_m != "muslim country") ~ 3,
     (is.na(religion_now) &  is.na(religion_before) &
       mbrncnt_m == "muslim country" & fbrncnt_m != "muslim country") ~ 3,
     
     (religion_now == "no religion" &  religion_before == "no religion" &
       fbrncnt_m == "muslim country" & mbrncnt_m != "muslim country") ~ 3,
     (religion_now == "no religion" &  is.na(religion_before) &
       fbrncnt_m == "muslim country" & mbrncnt_m != "muslim country") ~ 3,
     (is.na(religion_now) &  is.na(religion_before) &
       fbrncnt_m == "muslim country" & mbrncnt_m != "muslim country") ~ 3,
      
      
      (religion_now == "no religion" & is.na(religion_before) & is.na(cntbrth_m) & is.na(fbrncnt_m) & is.na(mbrncnt_m)) ~ 0,
      
      (is.na(religion_now) & is.na(religion_before) & is.na(cntbrth_m) & is.na(fbrncnt_m) & is.na(mbrncnt_m)) ~ 99,
      ))%>%
  mutate(
    dikotom = case_when(
     respondent %in% c(1,2) ~ "muslim",
     
     (religion_now %in% c("no religion", "not muslim") & religion_before != "muslim" & is.na(cntbrth_m)) ~ "majority",
     (religion_now %in% c("no religion", "not muslim") & is.na(religion_before) & is.na(cntbrth_m)) ~ "majority"

    ))%>%
  mutate(dikotom_r = case_when(
    respondent %in% c(1) ~ "muslim",
    dikotom == "majority" ~ "majority"
  ))

muslim$respondent[is.na(muslim$respondent)] <- 999
```




```{r}
muslim <- muslim %>%
  mutate(
    step_1 = case_when(
      (religion_now == "muslim") ~ 1, TRUE ~ 0
      ),
    
    step_2 = case_when(
      (religion_now == "no religion" &  religion_before == "muslim") ~ 1, TRUE ~ 0
      ),
    
    step_3 = case_when(
      (religion_now == "no religion" &  religion_before == "no religion" & cntbrth_m == "muslim country") ~ 1,
      (religion_now == "no religion" &  is.na(religion_before) & cntbrth_m == "muslim country") ~ 1,
      (is.na(religion_now) &  is.na(religion_before) & cntbrth_m == "muslim country") ~ 1, TRUE ~ 0
      ),
    
    step_4 = case_when(
     (religion_now == "no religion" &  religion_before == "no religion" &
       mbrncnt_m == "muslim country" & fbrncnt_m == "muslim country") ~ 1,
     (religion_now == "no religion" &  is.na(religion_before) &
       mbrncnt_m == "muslim country" & fbrncnt_m == "muslim country") ~ 1,
     (is.na(religion_now) &  is.na(religion_before) &
       mbrncnt_m == "muslim country" & fbrncnt_m == "muslim country") ~ 1, TRUE ~ 0
    ),
     step_4l = case_when(
     respondent == 3 ~ 1, TRUE ~ 0)
    
    )%>%
  mutate(step_12 = case_when(
    step_1 == 1 | step_2 == 1 ~ 1, TRUE ~ 0
  ))%>%
  mutate(step_123 = case_when(
    step_12 == 1 | step_3 == 1 ~ 1, TRUE ~ 0
  ))%>%
  mutate(step_1234 = case_when(
    step_123 == 1 | step_4 == 1 ~ 1, TRUE ~ 0
  ))%>%
  mutate(step_1234l = case_when(
    step_1234 == 1 | step_4l == 1 ~ 1, TRUE ~ 0
  ))
```

```{r}
tabl <- function(...) table(..., useNA='ifany')
```

```{r}
#tabl(muslim$respondent)
tabl(ess_immigration$rlgdnm)
tabl(ess_immigration$rlgblg)

```

passer med ca. 4 procent muslim i EU - så ikke over sample

```{r}
tabl(muslim$dikotom)
tabl(muslim$step_4l)
tabl(muslim$dikotom_r)
```

```{r}
muslim <- muslim %>%
  dplyr::select(id, respondent, dikotom, step_1:step_1234l, dikotom_r)

ess_immigration <- left_join(ess_immigration, muslim, by = "id")
```

# Gemme datasæt

```{r}
tabl <- function(...) table(..., useNA='ifany')
tabl(ess_immigration$mnrgtjb)
tabl(ess_immigration$hmsfmlsh)
tabl(ess_immigration$hmsacld)
tabl(ess_immigration$rlgatnd)
tabl(ess_immigration$rlgdgr)
```

```{r}
ess_immigration <- ess_immigration%>%
  filter(cntry %in% c("NO", "SE", "DK", "DE", "GB", "IT", "ES", "NL", "AT", "FR", "PT", "FI", "GR", "CH", "IE", "BE", "LU", "IS"))
```


```{r}
write.csv(ess_immigration, "../data/ess_treated.csv")
end.time <- Sys.time()
time <- end.time - start.time
time
```

```{r}
tabl(ess_immigration$dikotom, ess_immigration$cntry)
```
